feat: migrated from create-react-app to rollup bundler
