+++ lib/index.ts
@@ -1,0 +1,3 @@
+import useOtp from "./hook/useOtp";
+
+export { useOtp };
