+++ rollup.config.js
@@ -1,0 +1,22 @@
+import peerDepsExternal from "rollup-plugin-peer-deps-external";
+import resolve from "@rollup/plugin-node-resolve";
+import typescript from "rollup-plugin-typescript2";
+
+const packageJson = require("./package.json");
+
+export default {
+  input: "lib/index.ts",
+  output: [
+    {
+      file: packageJson.module,
+      format: "esm",
+      sourcemap: true,
+    },
+  ],
+  plugins: [
+    peerDepsExternal(),
+    resolve(),
+    typescript({ useTsconfigDeclarationDir: true }),
+  ],
+  external: ["react", "react-dom"]
+};
