--- src/App.tsx
@@ -1,16 +1,0 @@
-import React from 'react';
-import logo from './logo.svg';
-import './App.css';
-import Otp from './components/Otp';
-
-function App() {
-  return (
-    <div className="App">
-      <div>
-        <Otp />
-      </div>
-    </div>
-  );
-}
-
-export default App;
